+++ clients/fides-js/src/lib/tcf.ts
@@ -161,25 +161,22 @@ export const generateFidesString = async ({
         }
       });
 
-      // Set legitimate interest for special-purpose only vendors
+      // Set legitimate interest for special-purpose vendors
       if (experience.gvl?.vendors) {
         (experience as PrivacyExperience).tcf_vendor_relationships?.forEach(
           (relationship) => {
             const { id } = decodeVendorId(relationship.id);
             const vendor = experience.gvl?.vendors[id];
-            const isInVendorConsents = (
-              experience as PrivacyExperience
-            ).tcf_vendor_consents?.some(
-              (consent) => consent.id === relationship.id,
-            );
+            // ensure specialPurposes includes purposes that are not forbidden
+            const nonForbiddenSpecialPurposes =
+              vendor?.specialPurposes?.filter(
+                (p) => !FORBIDDEN_LEGITIMATE_INTEREST_PURPOSE_IDS.includes(p),
+              ) || [];
             if (
               vendor &&
-              vendor.specialPurposes?.length &&
-              (!vendor.purposes || vendor.purposes.length === 0) &&
-              (!vendor.flexiblePurposes ||
-                vendor.flexiblePurposes.length === 0) &&
-              (!vendor.legIntPurposes || vendor.legIntPurposes.length === 0) &&
-              !isInVendorConsents
+              (vendor.purposes.length === 0 || vendor.purposes.length > 0) &&
+              vendor.legIntPurposes.length === 0 &&
+              nonForbiddenSpecialPurposes.length > 0
             ) {
               tcModel.vendorLegitimateInterests.set(+id);
             }